+++ packages/assets-controllers/src/TokensController.ts
@@ -8,11 +8,10 @@ import type {
 } from '@metamask/accounts-controller';
 import type { AddApprovalRequest } from '@metamask/approval-controller';
 import type {
-  RestrictedMessenger,
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-} from '@metamask/base-controller';
-import { BaseController } from '@metamask/base-controller';
+} from '@metamask/base-controller/next';
+import { BaseController } from '@metamask/base-controller/next';
 import contractsMap from '@metamask/contract-metadata';
 import {
   toChecksumHexAddress,
@@ -27,6 +26,7 @@ import {
 } from '@metamask/controller-utils';
 import type { KeyringControllerAccountRemovedEvent } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
+import type { Messenger } from '@metamask/messenger';
 import { abiERC721 } from '@metamask/metamask-eth-abis';
 import type {
   NetworkClientId,
@@ -152,12 +152,10 @@ export type AllowedEvents =
 /**
  * The messenger of the {@link TokensController}.
  */
-export type TokensControllerMessenger = RestrictedMessenger<
+export type TokensControllerMessenger = Messenger<
   typeof controllerName,
   TokensControllerActions | AllowedActions,
-  TokensControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  TokensControllerEvents | AllowedEvents
 >;
 
 export const getDefaultTokensState = (): TokensControllerState => {
@@ -218,32 +216,32 @@ export class TokensController extends BaseController<
 
     this.#abortController = new AbortController();
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:addDetectedTokens` as const,
       this.addDetectedTokens.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:addTokens` as const,
       this.addTokens.bind(this),
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'AccountsController:selectedEvmAccountChange',
       this.#onSelectedAccountChange.bind(this),
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'NetworkController:stateChange',
       this.#onNetworkStateChange.bind(this),
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'KeyringController:accountRemoved',
       (accountAddress: string) => this.#handleOnAccountRemoved(accountAddress),
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TokenListController:stateChange',
       ({ tokensChainsCache }) => {
         const { allTokens } = this.state;
@@ -413,7 +411,7 @@ export class TokensController extends BaseController<
     const releaseLock = await this.#mutex.acquire();
     const { allTokens, allIgnoredTokens, allDetectedTokens } = this.state;
 
-    const chainIdToUse = this.messagingSystem.call(
+    const chainIdToUse = this.messenger.call(
       'NetworkController:getNetworkClientById',
       networkClientId,
     ).configuration.chainId;
@@ -499,7 +497,7 @@ export class TokensController extends BaseController<
     const { allTokens, allIgnoredTokens, allDetectedTokens } = this.state;
     const importedTokensMap: { [key: string]: true } = {};
 
-    const interactingChainId = this.messagingSystem.call(
+    const interactingChainId = this.messenger.call(
       'NetworkController:getNetworkClientById',
       networkClientId,
     ).configuration.chainId;
@@ -575,7 +573,7 @@ export class TokensController extends BaseController<
     tokenAddressesToIgnore: string[],
     networkClientId: NetworkClientId,
   ) {
-    const interactingChainId = this.messagingSystem.call(
+    const interactingChainId = this.messenger.call(
       'NetworkController:getNetworkClientById',
       networkClientId,
     ).configuration.chainId;
@@ -733,7 +731,7 @@ export class TokensController extends BaseController<
     tokenAddress: string,
     networkClientId: NetworkClientId,
   ) {
-    const chainIdToUse = this.messagingSystem.call(
+    const chainIdToUse = this.messenger.call(
       'NetworkController:getNetworkClientById',
       networkClientId,
     ).configuration.chainId;
@@ -792,7 +790,7 @@ export class TokensController extends BaseController<
   #getProvider(networkClientId?: NetworkClientId): Web3Provider {
     return new Web3Provider(
       networkClientId
-        ? this.messagingSystem.call(
+        ? this.messenger.call(
             'NetworkController:getNetworkClientById',
             networkClientId,
           ).provider
@@ -1076,7 +1074,7 @@ export class TokensController extends BaseController<
   }
 
   async #requestApproval(suggestedAssetMeta: SuggestedAssetMeta) {
-    return this.messagingSystem.call(
+    return this.messenger.call(
       'ApprovalController:addRequest',
       {
         id: suggestedAssetMeta.id,
@@ -1098,12 +1096,12 @@ export class TokensController extends BaseController<
   }
 
   #getSelectedAccount() {
-    return this.messagingSystem.call('AccountsController:getSelectedAccount');
+    return this.messenger.call('AccountsController:getSelectedAccount');
   }
 
   #getSelectedAddress() {
     // If the address is not defined (or empty), we fallback to the currently selected account's address
-    const account = this.messagingSystem.call(
+    const account = this.messenger.call(
       'AccountsController:getAccount',
       this.#selectedAccountId,
     );