+++ packages/transaction-controller/src/TransactionController.ts
@@ -11,9 +11,8 @@ import type {
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-  RestrictedMessenger,
-} from '@metamask/base-controller';
-import { BaseController } from '@metamask/base-controller';
+} from '@metamask/base-controller/next';
+import { BaseController } from '@metamask/base-controller/next';
 import {
   query,
   ApprovalType,
@@ -27,6 +26,7 @@ import type {
   GasFeeState,
 } from '@metamask/gas-fee-controller';
 import type { KeyringControllerSignEip7702AuthorizationAction } from '@metamask/keyring-controller';
+import type { Messenger } from '@metamask/messenger';
 import type {
   BlockTracker,
   NetworkClientId,
@@ -671,12 +671,10 @@ export type TransactionControllerEvents =
 /**
  * The messenger of the {@link TransactionController}.
  */
-export type TransactionControllerMessenger = RestrictedMessenger<
+export type TransactionControllerMessenger = Messenger<
   typeof controllerName,
   TransactionControllerActions | AllowedActions,
-  TransactionControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  TransactionControllerEvents | AllowedEvents
 >;
 
 /**
@@ -868,7 +866,7 @@ export class TransactionController extends BaseController<
       },
     });
 
-    this.messagingSystem = messenger;
+    this.messenger = messenger;
 
     this.#afterAdd = hooks?.afterAdd ?? (() => Promise.resolve({}));
     this.#afterSign = hooks?.afterSign ?? (() => true);
@@ -916,7 +914,7 @@ export class TransactionController extends BaseController<
     this.#transactionHistoryLimit = transactionHistoryLimit;
 
     const findNetworkClientIdByChainId = (chainId: Hex) => {
-      return this.messagingSystem.call(
+      return this.messenger.call(
         `NetworkController:findNetworkClientIdByChainId`,
         chainId,
       );
@@ -925,7 +923,7 @@ export class TransactionController extends BaseController<
     this.#multichainTrackingHelper = new MultichainTrackingHelper({
       findNetworkClientIdByChainId,
       getNetworkClientById: ((networkClientId: NetworkClientId) => {
-        return this.messagingSystem.call(
+        return this.messenger.call(
           `NetworkController:getNetworkClientById`,
           networkClientId,
         );
@@ -937,10 +935,7 @@ export class TransactionController extends BaseController<
       createPendingTransactionTracker:
         this.#createPendingTransactionTracker.bind(this),
       onNetworkStateChange: (listener) => {
-        this.messagingSystem.subscribe(
-          'NetworkController:stateChange',
-          listener,
-        );
+        this.messenger.subscribe('NetworkController:stateChange', listener);
       },
     });
 
@@ -956,12 +951,9 @@ export class TransactionController extends BaseController<
       getTransactions: () => this.state.transactions,
       getTransactionBatches: () => this.state.transactionBatches,
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       onStateChange: (listener) => {
-        this.messagingSystem.subscribe(
-          'TransactionController:stateChange',
-          listener,
-        );
+        this.messenger.subscribe('TransactionController:stateChange', listener);
       },
     });
 
@@ -996,7 +988,7 @@ export class TransactionController extends BaseController<
       includeTokenTransfers:
         this.#incomingTransactionOptions.includeTokenTransfers,
       isEnabled: this.#incomingTransactionOptions.isEnabled,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       remoteTransactionSource: new AccountsApiRemoteTransactionSource(),
       trimTransactions: this.#trimTransactionsForState.bind(this),
       updateTransactions: this.#incomingTransactionOptions.updateTransactions,
@@ -1008,7 +1000,7 @@ export class TransactionController extends BaseController<
 
     // when transactionsController state changes
     // check for pending transactions and start polling if there are any
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TransactionController:stateChange',
       this.#checkForPendingTransactionAndStartPolling,
     );
@@ -1016,7 +1008,7 @@ export class TransactionController extends BaseController<
     new ResimulateHelper({
       simulateTransaction: this.#updateSimulationData.bind(this),
       onTransactionsUpdate: (listener) => {
-        this.messagingSystem.subscribe(
+        this.messenger.subscribe(
           'TransactionController:stateChange',
           listener,
           (controllerState) => controllerState.transactions,
@@ -1060,7 +1052,7 @@ export class TransactionController extends BaseController<
   async addTransactionBatch(
     request: TransactionBatchRequest,
   ): Promise<TransactionBatchResult> {
-    const { blockTracker } = this.messagingSystem.call(
+    const { blockTracker } = this.messenger.call(
       `NetworkController:getNetworkClientById`,
       request.networkClientId,
     );
@@ -1082,7 +1074,7 @@ export class TransactionController extends BaseController<
       getTransaction: (transactionId) =>
         this.#getTransactionOrThrow(transactionId),
       isSimulationEnabled: this.#isSimulationEnabled,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       publishBatchHook: this.#publishBatchHook,
       publicKeyEIP7702: this.#publicKeyEIP7702,
       publishTransaction: (
@@ -1107,7 +1099,7 @@ export class TransactionController extends BaseController<
     return isAtomicBatchSupported({
       ...request,
       getEthQuery: (chainId) => this.#getEthQuery({ chainId }),
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       publicKeyEIP7702: this.#publicKeyEIP7702,
     });
   }
@@ -1335,7 +1327,7 @@ export class TransactionController extends BaseController<
         {
           isSwapsDisabled: this.#isSwapsDisabled,
           cancelTransaction: this.#rejectTransaction.bind(this),
-          messenger: this.messagingSystem,
+          messenger: this.messenger,
         },
       );
 
@@ -1360,7 +1352,7 @@ export class TransactionController extends BaseController<
         );
       }
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:unapprovedTransactionAdded`,
         addedTransactionMeta,
       );
@@ -1428,7 +1420,7 @@ export class TransactionController extends BaseController<
         txParams.value = '0x0';
       },
       afterSubmit: (newTransactionMeta) => {
-        this.messagingSystem.publish(
+        this.messenger.publish(
           `${controllerName}:transactionFinished`,
           newTransactionMeta,
         );
@@ -1467,7 +1459,7 @@ export class TransactionController extends BaseController<
       transactionId,
       transactionType: TransactionType.retry,
       afterSubmit: (newTransactionMeta) => {
-        this.messagingSystem.publish(
+        this.messenger.publish(
           `${controllerName}:speedupTransactionAdded`,
           newTransactionMeta,
         );
@@ -1583,12 +1575,12 @@ export class TransactionController extends BaseController<
     this.#addMetadata(newTransactionMeta);
 
     // speedUpTransaction has no approval request, so we assume the user has already approved the transaction
-    this.messagingSystem.publish(`${controllerName}:transactionApproved`, {
+    this.messenger.publish(`${controllerName}:transactionApproved`, {
       transactionMeta: newTransactionMeta,
       actionId,
     });
 
-    this.messagingSystem.publish(`${controllerName}:transactionSubmitted`, {
+    this.messenger.publish(`${controllerName}:transactionSubmitted`, {
       transactionMeta: newTransactionMeta,
       actionId,
     });
@@ -1624,7 +1616,7 @@ export class TransactionController extends BaseController<
       ignoreDelegationSignatures,
       isSimulationEnabled: this.#isSimulationEnabled(),
       getSimulationConfig: this.#getSimulationConfig,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       txParams: transaction,
     });
 
@@ -1653,7 +1645,7 @@ export class TransactionController extends BaseController<
       ethQuery,
       isSimulationEnabled: this.#isSimulationEnabled(),
       getSimulationConfig: this.#getSimulationConfig,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       txParams: transaction,
     });
 
@@ -1800,7 +1792,7 @@ export class TransactionController extends BaseController<
         throw error;
       });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:transactionConfirmed`,
         updatedTransactionMeta,
       );
@@ -2144,7 +2136,7 @@ export class TransactionController extends BaseController<
 
     await updateTransactionLayer1GasFee({
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       provider,
       transactionMeta: updatedTransaction,
     });
@@ -2338,7 +2330,7 @@ export class TransactionController extends BaseController<
         status as TransactionStatus,
       )
     ) {
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:transactionFinished`,
         updatedTransactionMeta,
       );
@@ -2472,7 +2464,7 @@ export class TransactionController extends BaseController<
     const gasFeeFlow = getGasFeeFlow(
       transactionMeta,
       this.#gasFeeFlows,
-      this.messagingSystem,
+      this.messenger,
     ) as GasFeeFlow;
 
     const ethQuery = new EthQuery(provider);
@@ -2484,7 +2476,7 @@ export class TransactionController extends BaseController<
     return gasFeeFlow.getGasFees({
       ethQuery,
       gasFeeControllerData,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       transactionMeta,
     });
   }
@@ -2514,7 +2506,7 @@ export class TransactionController extends BaseController<
 
     return await getTransactionLayer1GasFee({
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       provider,
       transactionMeta: {
         txParams: transactionParams,
@@ -2790,7 +2782,7 @@ export class TransactionController extends BaseController<
           gasFeeFlows: this.#gasFeeFlows,
           getGasFeeEstimates: this.#getGasFeeEstimates,
           getSavedGasFees: this.#getSavedGasFees.bind(this),
-          messenger: this.messagingSystem,
+          messenger: this.messenger,
           txMeta: transactionMeta,
         }),
     );
@@ -2800,7 +2792,7 @@ export class TransactionController extends BaseController<
       async () =>
         await updateTransactionLayer1GasFee({
           layer1GasFeeFlows: this.#layer1GasFeeFlows,
-          messenger: this.messagingSystem,
+          messenger: this.messenger,
           provider,
           transactionMeta,
         }),
@@ -2907,13 +2899,10 @@ export class TransactionController extends BaseController<
           const updatedTransactionMeta = this.#getTransaction(
             transactionId,
           ) as TransactionMeta;
-          this.messagingSystem.publish(
-            `${controllerName}:transactionApproved`,
-            {
-              transactionMeta: updatedTransactionMeta,
-              actionId,
-            },
-          );
+          this.messenger.publish(`${controllerName}:transactionApproved`, {
+            transactionMeta: updatedTransactionMeta,
+            actionId,
+          });
         }
       } catch (rawError: unknown) {
         const error = rawError as Error & { code?: number; data?: Json };
@@ -3045,7 +3034,7 @@ export class TransactionController extends BaseController<
 
       if (!(await this.#beforePublish(transactionMeta))) {
         log('Skipping publishing transaction based on hook');
-        this.messagingSystem.publish(
+        this.messenger.publish(
           `${controllerName}:transactionPublishingSkipped`,
           transactionMeta,
         );
@@ -3125,11 +3114,11 @@ export class TransactionController extends BaseController<
         },
       );
 
-      this.messagingSystem.publish(`${controllerName}:transactionSubmitted`, {
+      this.messenger.publish(`${controllerName}:transactionSubmitted`, {
         transactionMeta,
       });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:transactionFinished`,
         transactionMeta,
       );
@@ -3187,7 +3176,7 @@ export class TransactionController extends BaseController<
       error: normalizeTxError(error ?? providerErrors.userRejectedRequest()),
     };
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:transactionFinished`,
       updatedTransactionMeta,
     );
@@ -3197,7 +3186,7 @@ export class TransactionController extends BaseController<
       updatedTransactionMeta,
     );
 
-    this.messagingSystem.publish(`${controllerName}:transactionRejected`, {
+    this.messenger.publish(`${controllerName}:transactionRejected`, {
       transactionMeta: updatedTransactionMeta,
       actionId,
     });
@@ -3299,7 +3288,7 @@ export class TransactionController extends BaseController<
       parentContext: traceContext,
     });
 
-    return (await this.messagingSystem.call(
+    return (await this.messenger.call(
       'ApprovalController:addRequest',
       {
         id,
@@ -3435,7 +3424,7 @@ export class TransactionController extends BaseController<
       );
     });
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:incomingTransactionsReceived`,
       finalTransactions,
     );
@@ -3584,7 +3573,7 @@ export class TransactionController extends BaseController<
       ...transactionMeta,
       status: TransactionStatus.dropped as const,
     };
-    this.messagingSystem.publish(`${controllerName}:transactionDropped`, {
+    this.messenger.publish(`${controllerName}:transactionDropped`, {
       transactionMeta: updatedTransactionMeta,
     });
     this.updateTransaction(
@@ -3674,7 +3663,7 @@ export class TransactionController extends BaseController<
 
     const signedAuthorizationList = await signAuthorizationList({
       authorizationList,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       transactionMeta,
     });
 
@@ -3766,7 +3755,7 @@ export class TransactionController extends BaseController<
   }
 
   #onTransactionStatusChange(transactionMeta: TransactionMeta) {
-    this.messagingSystem.publish(`${controllerName}:transactionStatusUpdated`, {
+    this.messenger.publish(`${controllerName}:transactionStatusUpdated`, {
       transactionMeta,
     });
   }
@@ -3789,7 +3778,7 @@ export class TransactionController extends BaseController<
 
     this.#markNonceDuplicatesDropped(transactionMeta.id);
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:transactionConfirmed`,
       transactionMeta,
     );
@@ -3820,7 +3809,7 @@ export class TransactionController extends BaseController<
           updateTransaction: this.updateTransaction.bind(this),
         });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:postTransactionBalanceUpdated`,
         {
           transactionMeta: updatedTransactionMeta,
@@ -3884,7 +3873,7 @@ export class TransactionController extends BaseController<
         this.#multichainTrackingHelper.acquireNonceLockForChainIdKey({
           chainId,
         }),
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       publishTransaction: (_ethQuery, transactionMeta) =>
         this.#publishTransaction(_ethQuery, transactionMeta, {
           skipSubmitHistory: true,
@@ -4236,7 +4225,7 @@ export class TransactionController extends BaseController<
         chainId,
         getSimulationConfig: this.#getSimulationConfig,
         isEIP7702GasFeeTokensEnabled: this.#isEIP7702GasFeeTokensEnabled,
-        messenger: this.messagingSystem,
+        messenger: this.messenger,
         publicKeyEIP7702: this.#publicKeyEIP7702,
         transactionMeta,
       });
@@ -4335,11 +4324,11 @@ export class TransactionController extends BaseController<
   }
 
   #getSelectedAccount() {
-    return this.messagingSystem.call('AccountsController:getSelectedAccount');
+    return this.messenger.call('AccountsController:getSelectedAccount');
   }
 
   #getInternalAccounts(): Hex[] {
-    const state = this.messagingSystem.call('AccountsController:getState');
+    const state = this.messenger.call('AccountsController:getState');
 
     return Object.values(state.internalAccounts?.accounts ?? {})
       .filter((account) => account.type === 'eip155:eoa')
@@ -4399,18 +4388,18 @@ export class TransactionController extends BaseController<
       isCustomNetwork,
       isSimulationEnabled: this.#isSimulationEnabled(),
       getSimulationConfig: this.#getSimulationConfig,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       txMeta: transactionMeta,
     });
   }
 
   #registerActionHandlers(): void {
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:estimateGas`,
       this.estimateGas.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:updateCustodialTransaction`,
       this.updateCustodialTransaction.bind(this),
     );
@@ -4484,7 +4473,7 @@ export class TransactionController extends BaseController<
       };
     }
 
-    this.messagingSystem.publish(`${controllerName}:transactionFailed`, {
+    this.messenger.publish(`${controllerName}:transactionFailed`, {
       actionId,
       error: error.message,
       transactionMeta: newTransactionMeta,
@@ -4492,7 +4481,7 @@ export class TransactionController extends BaseController<
 
     this.#onTransactionStatusChange(newTransactionMeta);
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:transactionFinished`,
       newTransactionMeta,
     );